Make unit testing works
