+++ src/lib/sqlite/global.ts
@@ -2,4 +2,4 @@ import { createDb, createQueryExecutor, createSQLiteAPI } from './wa-sqlite'
 
 export const sqlite3 = await createSQLiteAPI()
 export const executor = await createQueryExecutor(sqlite3, 'crypta')
-export const localDb = await createDb(executor)
+export const localDb = createDb(executor)
