+++ src/lib/sqlite/wa-sqlite.test.ts
@@ -3,10 +3,13 @@ import { setupServer } from 'msw/node'
 import {
   createDb,
   createSQLiteAPI,
-  createQueryExecutor,
+  createQueryExecutor, migrate,
 } from './wa-sqlite'
+import type { WASQLiteExecutor } from './wa-sqlite'
+import type { SqliteRemoteDatabase } from 'drizzle-orm/sqlite-proxy';
 import { createWASqliteMockWASMHandler } from '$lib/utitlities/testing'
 import { sql } from 'drizzle-orm'
+import { defaultMigrationQueryMap, defaultQueriesStringMap } from '$lib/sqlite/migration'
 
 const handlers = createWASqliteMockWASMHandler()
 const server = setupServer(...handlers)
@@ -17,10 +20,18 @@ afterAll(() => {
   server.close()
 })
 
-describe('executor', async () => {
+describe('happy path', async () => {
+  let sqliteAPI: SQLiteAPI
+  let executor: WASQLiteExecutor
+  let localDb: SqliteRemoteDatabase
+
+  beforeAll(async () => {
+    sqliteAPI = await createSQLiteAPI('http://mock.local', 'MemoryVFS')
+    executor = await createQueryExecutor(sqliteAPI, 'crypta', false)
+    localDb = createDb(executor)
+  })
+
   it('basic query', async () => {
-    const sqliteAPI = await createSQLiteAPI('http://mock.local', 'MemoryVFS')
-    const executor = await createQueryExecutor(sqliteAPI, 'crypta', false)
     {
       const result = await executor.execute('SELECT 1')
       expect(result).toEqual([[1]])
@@ -32,9 +43,6 @@ describe('executor', async () => {
   })
 
   it('local db', async () => {
-    const sqliteAPI = await createSQLiteAPI('http://mock.local', 'MemoryVFS')
-    const executor = await createQueryExecutor(sqliteAPI, 'crypta', false)
-    const localDb = createDb(executor)
     {
       const result = await localDb.run(sql`SELECT 1`)
       expect(result).toEqual({rows: [[1]]})
@@ -42,6 +50,19 @@ describe('executor', async () => {
     {
       const result = await localDb.run(sql`PRAGMA user_version`)
       expect(result).toEqual({rows: [[0]]})
+    }
+  })
+
+  it('migrate', async () => {
+    await migrate(
+      localDb,
+      async () => {},
+      defaultMigrationQueryMap,
+      defaultQueriesStringMap,
+    )
+    {
+      const result = await localDb.all(sql`SELECT * FROM folders`)
+      expect(result.length).toBe(1)
     }
   })
 })
